requirements.txt
langchain==0.3.14
langchain-community==0.3.14
langchain-google-genai==2.0.8
python-dotenv==1.0.1
streamlit==1.41.1
google-cloud-bigquery==3.27.0
google-cloud-storage==2.19.0
google-auth-oauthlib==1.2.1
google-analytics-data==0.18.16
chromadb==0.6.2